[GS/StackClash][WIP] Work towards checking the invariants based on https://developers.redhat.com/blog/2020/05/22/stack-clash-mitigation-in-gcc-part-3#
This basically boils down to checking that:
1. The stack grows at most 1 page at a time.
2. At every time, the closest-to-top-of-stack access needs to be at most within the 2 pages on the top of the stack.

To achieve this analysis, the StackClashAnalysis now tracks the guaranteed maximum offset from the current top-of-stack of the closest-to-top-of-stack access.

The current implementation is not good enough yet to not produce false positives on the gcc stack-clash protection on a very simple unknown-at-compile-time alloca case, but it is close.

It seems the one thing really missing is the analyzer recognizing the epilogue pattern
where the stack pointer gets the value of the frame pointer.